CAN remastered totally
CAN got an abstraction class called CanInterface now.
CAN bus can be accessed by means of the CanInterface or by using of CanSocket.
CanSocket is inherited from Device, so one can write and read to the CAN bus in the usual way. The main difference is that one must provide CAN ID and receive it in the first 2 or 4 bytes of data, dependding on the ExtId flag when the socket is opened.
Old Objnet::CanInterface renamed to CanOnbInterface and was rewritten too